+++ saga-annotation-sample/src/main/java/org/apache/seata/action/impl/SagaActionTwoImpl.java
@@ -16,9 +16,11 @@
  */
 package org.apache.seata.action.impl;
 
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
+
 import org.apache.seata.action.ResultHolder;
-import org.apache.seata.action.TccActionOne;
+import org.apache.seata.action.SagaActionTwo;
 import org.apache.seata.rm.tcc.api.BusinessActionContext;
 import org.apache.seata.rm.tcc.api.BusinessActionContextParameter;
 import org.apache.seata.rm.tcc.api.LocalTCC;
@@ -28,22 +30,31 @@ import org.springframework.util.Assert;
 
 @Service
 @LocalTCC
-public class TccActionOneImpl implements TccActionOne {
+public class SagaActionTwoImpl implements SagaActionTwo {
 
     @Override
-    @CompensationBusinessAction(name = "DubboSagaActionOne", compensationMethod = "rollback")
-    public boolean commit(BusinessActionContext actionContext,@BusinessActionContextParameter(paramName = "a") int a) {
+    @CompensationBusinessAction(name = "DubboTccActionTwo", compensationMethod = "rollback")
+    public boolean commit(BusinessActionContext actionContext,  @BusinessActionContextParameter(paramName = "b") String b,
+                           @BusinessActionContextParameter(paramName = "c", index = 1) List list) {
         String xid = actionContext.getXid();
-        System.out.println("TccActionOne prepare, xid:" + xid + ", a:" + a);
+        System.out.println("TccActionTwo prepare, xid:" + xid + ", b:" + b + ", c:" + list.get(1));
+        CompletableFuture<Void> completableFuture = new CompletableFuture<>();
+        ResultHolder.setActionTwoResult(xid,completableFuture);
+        completableFuture.complete(null);
         return true;
     }
 
     @Override
     public boolean rollback(BusinessActionContext actionContext) {
         String xid = actionContext.getXid();
-        Object a =  actionContext.getActionContext("a");
-        Assert.notNull(a);
-        System.out.println("TccActionOne prepare, xid:" + xid + ", a:" + a);
+        Assert.isTrue(actionContext.getActionContext("b") != null);
+        Assert.isTrue(actionContext.getActionContext("c") != null);
+        System.out.println(
+            "TccActionTwo rollback, xid:" + xid + ", b:" + actionContext.getActionContext("b") + ", c:" + actionContext
+                .getActionContext("c"));
+        CompletableFuture<Void> completableFuture = ResultHolder.getActionTwoResult(xid);
+        completableFuture.complete(null);
         return true;
     }
+
 }
